+++ packages/sdk/src/lib/types/coreTypes.ts
@@ -73,6 +73,14 @@ export type ProtectDataParams = {
   allowDebug?: boolean;
 
   /**
+   * specify the platform used for storing the encrypted payload of the protected data
+   *
+   * - `"ipfs"` (default): https://ipfs.tech/
+   * - `"arweave"`: https://arweave.org/
+   */
+  uploadMode?: 'ipfs' | 'arweave';
+
+  /**
    * Callback function that will get called at each step of the process
    */
   onStatusUpdate?: OnStatusUpdateFn<ProtectDataStatuses>;
@@ -98,6 +106,7 @@ export type ProtectedDataCreationProps = {
   transactionHash: string;
   zipFile: Uint8Array;
   encryptionKey: string;
+  multiaddr: string;
 };
 
 export type ProtectedDataWithSecretProps = ProtectedData &
@@ -276,6 +285,7 @@ export type TransferResponse = {
 
 // ---------------------ProcessProtectedData Types------------------------------------
 export type ProcessProtectedDataStatuses =
+  | 'FETCH_ORDERS'
   | 'FETCH_PROTECTED_DATA_ORDERBOOK'
   | 'FETCH_APP_ORDERBOOK'
   | 'FETCH_WORKERPOOL_ORDERBOOK'
@@ -302,11 +312,22 @@ export type ProcessProtectedDataParams = {
   userWhitelist?: Address;
 
   /**
-   * The maximum price per task for processing the protected data.
-   * It is the sum of the application price, dataset price and workerpool price per task.
+   * The maximum price of dataset per task for processing the protected data.
   @default = 0
   */
-  maxPrice?: number;
+  dataMaxPrice?: number;
+
+  /**
+   * The maximum price of application per task for processing the protected data.
+  @default = 0
+  */
+  appMaxPrice?: number;
+
+  /**
+   * The maximum price of workerpool per task for processing the protected data.
+  @default = 0
+  */
+  workerpoolMaxPrice?: number;
 
   /**
    * The file name of the desired file in the returned ZIP file.