+++ packages/sharing-smart-contract/config/env.js
@@ -1,0 +1,3 @@
+import env from './env.cjs';
+
+export default env;